buildDateTime + calculateTotals helpers
