+++ l2-frontend/src/pages/WorkingTime/DateCell.vue
@@ -130,6 +130,7 @@ import RadioFieldById from '@/fields/RadioFieldById.vue';
 import {
   ShiftVariantItem, TimeOptionItem, WorkDayStatusItem, WorkTimeDayCell,
 } from '@/pages/WorkingTime/types/types';
+import { buildDateTime } from '@/pages/WorkingTime/utils/date';
 
 const emit = defineEmits(['changeWorkTime', 'copyPrevFilled', 'copyToColumn']);
 const props = defineProps({
@@ -249,7 +250,7 @@ const timeOff = () => {
 
 const updateTime = async () => {
   if ((startWork.value && endWork.value) && endWork.value <= startWork.value && endWork.value !== '00:00') {
-    const start = new Date(`${props.date} ${startWork.value}`);
+    const start = buildDateTime(props.date, startWork.value);
     const endTime = endWork.value.split(':');
     nextDayEndWork.value = new Date(start.getFullYear(), start.getMonth(), start.getDate() + 1, endTime[0], endTime[1]);
     endWork.value = '00:00';
@@ -291,7 +292,7 @@ const selectedShift = ref(null);
 
 watch(selectedShift, () => {
   if (selectedShift.value) {
-    const start = new Date(`${props.date} ${startWork.value}`);
+    const start = buildDateTime(props.date, startWork.value);
     const shiftInMinutes = Number(selectedShift.value) * 60;
     const shiftWithLunch = shiftInMinutes + props.lunchDuration;
     const end = new Date(start.getTime() + (shiftWithLunch * 60 * 1000));
