+++ l2-frontend/src/pages/WorkingTime/TemplateTable.vue
@@ -22,7 +22,7 @@ import {
   EMPTY_WORK_TIME_DAY_CELL,
   HolidaysMap, ShiftVariantItem, TableColumn, TimeOptionItem, WorkDayStatusItem,
 } from '@/pages/WorkingTime/types/types';
-import {formatDateKey, formatDateTitle, isWeekend} from '@/pages/WorkingTime/utils/date';
+import { formatDateKey, formatDateTitle, isWeekend } from '@/pages/WorkingTime/utils/date';
 
 const props = defineProps({
   workDayStatuses: {
