+++ l2-frontend/src/pages/WorkingTime/utils/workTime.ts
@@ -1,0 +1,34 @@
+import { buildDateTime, isISODateString } from '@/pages/WorkingTime/utils/date';
+import { WorkTimeDayCell } from '@/pages/WorkingTime/types/types';
+
+const calculateEmployeeTotals = (employee) => {
+  let totalDiffTime = 0;
+  const keys = Object.keys(employee);
+  const lunchDuration = employee.lunchDuration * 60 * 1000;
+  for (const key of keys) {
+    if (isISODateString(key)) {
+      const currentDay: WorkTimeDayCell = employee[key];
+      if (currentDay.startWorkTime && currentDay.endWorkTime && !currentDay.typeId) {
+        const startTime = buildDateTime(key, currentDay.startWorkTime);
+        let endTime;
+        if (currentDay.endWorkTime === '00:00') {
+          endTime = new Date(startTime.getFullYear(), startTime.getMonth(), startTime.getDate() + 1, 0, 0);
+        } else {
+          endTime = buildDateTime(key, currentDay.endWorkTime);
+        }
+        const dayDiffTime = endTime.getTime() - startTime.getTime() - lunchDuration;
+        totalDiffTime += dayDiffTime;
+      }
+    }
+  }
+  const totalDiffSec = totalDiffTime / (1000 * 60);
+  const totalHoursDecimal = totalDiffSec / 60;
+  const totalHours = Math.trunc(totalHoursDecimal);
+  const totalMin = totalDiffSec % 60;
+  return {
+    totalHoursDecimal: totalHoursDecimal.toFixed(1),
+    totalHours: `${totalHours}ч ${totalMin}м`,
+  };
+};
+
+export default calculateEmployeeTotals;
